requirements.txt
chromadb==0.5.20
langchain==0.3.8
langchain_chroma==0.1.4
langchain_community==0.3.8
langchain_groq==0.2.1
langchain_openai==0.2.9
python-dotenv==1.0.1
